% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalMPLdc-package.R
\docType{package}
\name{survivalMPLdc-package}
\alias{survivalMPLdc-package}
\alias{survivalMPLdc}
\title{Penalised Likelihood for Survival Analysis with Dependent Censoring}
\description{
Penalised Likelihood for Survival Analysis with Dependent Censoring
}
\references{
Ma, J. (2010). \emph{"Positively constrained multiplicative iterative algorithm
for maximum penalised likelihood tomographic reconstruction"}. IEEE Transactions On Signal Processing
57, 181-192.

Brodaty H, Woodward M, Boundy K, Ames D, Balshaw R. (2011). \emph{"Patients in Australian memory clinics:
baseline characteristics and predictors of decline at six months"}. Int Psychogeriatr 23, 1086-1096.

Ma, J. and Heritier, S. and Lo, S. (2014). \emph{"On the Maximum Penalised Likelihood Approach
for Proportional Hazard Models with Right Censored Survival Data"}. Computational Statistics and
Data Analysis 74, 142-156.

Brodaty H, Connors M, Xu J, Woodward M, Ames D. (2014). \emph{"Predictors of institutionalization
in dementia: a three year longitudinal study"}. Journal of Alzheimers Disease 40, 221-226.

Xu J, Ma J, Connors MH, Brodaty H. (2018). \emph{"Proportional hazard model estimation under
dependent censoring using copulas and penalized likelihood"}. Statistics in Medicine 37, 2238–2251.
}
\keyword{package}
